Reorganize existing tests
